Rearrange packages.
